// src/types/StorageKeys.ts
export enum StorageKeys {
    ACCESS_TOKEN = 'api_access_token',
    REFRESH_TOKEN = 'api_refresh_token',
    LOGGED_USER_ID = 'api_logged_user_id',
    LOGGED_USER_USERNAME = 'api_logged_username',
    LOGGED_USER_FIRSTNAME = 'api_logged_firstname',
    LOGGED_USER_LASTNAME = 'api_logged_lastname',
}